[ADD] udes_common: Add CommonBase test class
Useful helper methods refactored out of TestGroupby, which now inherits
the new class

Story/2911
