+++ addons/udes_common/tests/__init__.py
@@ -1,3 +1,4 @@
+from . import common
 from . import test_groupby
 from . import test_iterators
 from . import test_selection_display_name
